MOSS and BattleEye Setup Script
I needed to make sure either BattleEye or MOSS is installed on the host or configured. Since I run a Cafe with Static RO images and don't want to keep the items installed, I wrote some scripts to make it easy to deploy them with the newest version available from their respective websites.

MOSS devs were able to provide me with a JSON file that I can pull for the latest supported games.